+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add Atw zone operation mode control.
+
+### Removed
+- Remove heat/cool simplification from Atw zone operation modes.
 
 ## [2.7.0] - 2020-05-01
 ### Added